Implemented count with Long value

+++ web/src/main/java/com/doggogram/backendsvc/controller/v1/UserController.java
@@ -44,7 +44,7 @@ public class UserController {
     }
 
     @GetMapping ({"/$count", "/$count/"})
-    public ResponseEntity<Integer> getCount() throws ControllerCountException {
+    public ResponseEntity<Long> getCount() throws ControllerCountException {
         try {
             return new ResponseEntity<>(userService.count(), HttpStatus.OK);
         } catch(Exception e) {
